refactor: extract cart types
