+++ packages/platform-core/src/cart/cartLine.d.ts
@@ -1,0 +1,6 @@
+import type { SKU } from "@acme/types";
+export interface CartLine {
+  sku: SKU;
+  qty: number;
+  size?: string;
+}
