+++ packages/platform-core/src/cart/cartLine.ts
@@ -1,0 +1,7 @@
+import type { SKU } from "@acme/types";
+
+export interface CartLine {
+  sku: SKU;
+  qty: number;
+  size?: string;
+}
