+++ packages/platform-core/src/cart/cartState.d.ts
@@ -1,0 +1,2 @@
+import type { CartLine } from "./cartLine";
+export type CartState = Record<string, CartLine>;
