+++ packages/platform-core/src/cart/cartState.ts
@@ -1,0 +1,3 @@
+import type { CartLine } from "./cartLine";
+
+export type CartState = Record<string, CartLine>;
