+++ packages/platform-core/src/cart/index.d.ts
@@ -1,0 +1,2 @@
+export type { CartLine } from "./cartLine";
+export type { CartState } from "./cartState";
