+++ packages/platform-core/src/cartStore/memoryStore.d.ts
@@ -1,4 +1,4 @@
-import type { CartState } from "../cartCookie";
+import type { CartState } from "../cart";
 import type { SKU } from "@acme/types";
 import type { CartStore } from "../cartStore";
 /** In-memory implementation of CartStore */
