+++ packages/platform-core/src/cartStore/redisStore.ts
@@ -1,7 +1,7 @@
 import crypto from "crypto";
 import type { Redis } from "@upstash/redis";
 
-import type { CartState } from "../cartCookie";
+import type { CartState } from "../cart";
 import type { SKU } from "@acme/types";
 import type { CartStore } from "../cartStore";
