+++ packages/platform-core/src/contexts/CartContext.d.ts
@@ -1,4 +1,4 @@
-import { type CartState } from "../cartCookie";
+import { type CartState } from "../cart";
 import type { SKU } from "@acme/types";
 import { type ReactNode } from "react";
 type Action = {
